A number of modifications to PerformFit.
Revised includes. Added an additional TF1 to estimate the parameters to
be used as initial input to assit convergence of fitting. Increased the
verbosity of the minuit fitter. Added an output of the fit results as
the output of Minuit may be incorrect if the fitting fails. Added
aditional output of the Minos error estimates. Fixed the graph axis such
that the plotting always starts from 0.